Creation of end to end audit script
This script, which makes use of the Enhanced Scripting Environment,
supports end to end auditing of a domains credentials.

The script will start by finding users using Get-ADUser, and this can
recieve either an identity or a filter+Searchbase+searchscope, The
script will then call the Find-ADUserPassword CMDLet.

The script can optional save what it found to a file, or send it via an
HTML email (specified separately from the Enhanced Script Environment).
This will by default include the passwords in plaintext.